New: Update AutoTags on series update
Closes #6783
